+++ Observer/StringExtensions.cs
@@ -11,12 +11,12 @@ namespace Observer
 		/// Computes a hash value for the string provided.  This is a simple algorithm, but it will remain stable
 		/// across .Net versions, something not guaranteed by the API otherwise.  (Although alternatives exist.)
 		/// </summary>
-		public static UInt64 ComputeHash(this string str)
+		public static UInt32 ComputeHash(this string str)
 		{
-			UInt64 hash = 42;
+			UInt32 hash = 42;
 			foreach (char c in str)
 			{
-				hash = ((hash << 5) + hash) + (UInt64) c;
+				hash = ((hash << 5) + hash) + (UInt32) c;
 			}
 
 			return hash;
